requirements.txt
cplex
Pillow
sv-ttk
ttkthemes
